requirements.txt
av==9.2.0
mediapipe==0.8.10.1
numpy==1.23.1
opencv_python_headless==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
requests==2.22.0
streamlit==1.10.0
streamlit_webrtc==0.42.0
